Создает редьюсер nearbyOffers
