+++ project/src/const.ts
@@ -83,5 +83,6 @@ export const REVIEW_RATE = [
 export enum NameSpaces {
   offers = 'OFFERS',
   reviews = 'REVIEWS',
-  utility = 'UTILITY'
+  utility = 'UTILITY',
+  nearbyOffers = 'NEARBY_OFFERS',
 }
