+++ project/src/store/api-actions.ts
@@ -4,7 +4,8 @@ import {api} from '../store';
 import {store} from '../store';
 import {Offer} from '../types/offer';
 import {redirectToRoute} from './action';
-import {loadOffers, loadOffer, loadNearbyOffers, loadFavoriteOffers, updateOfferFavoriteStatus, updateCurrentOffer, updateNearbyOffers} from './reducers/offers';
+import {loadOffers, loadOffer, loadFavoriteOffers, updateOfferFavoriteStatus, updateCurrentOffer} from './reducers/offers';
+import {loadNearbyOffers, updateNearbyOffers} from './reducers/nearby-offers';
 import {loadReviews, sendReviews} from './reducers/reviews';
 import {setError, requireAuthorization} from './reducers/utility';
 import {errorHandle} from '../sevrices/error-handle';
